requirements.txt
absl-py==1.4.0
array-record==0.5.0
asttokens==2.4.1
astunparse==1.6.3
blessed==1.20.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
dm-tree==0.1.8
editor==1.6.6
etils==1.7.0
executing==2.0.1
flatbuffers==24.3.7
fonttools==4.49.0
fsspec==2024.2.0
gast==0.4.0
google-auth==2.28.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.62.0
grpcio==1.62.0
h5py==3.10.0
idna==3.6
importlib_resources==6.1.3
inquirer==3.2.4
ipykernel==6.29.3
ipython==8.22.2
jedi==0.19.1
joblib==1.3.2
jupyter_client==8.6.0
jupyter_core==5.7.1
keras==2.13.1
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.5.2
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
nest-asyncio==1.6.0
numpy==1.24.3
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
pandas==2.2.1
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
prettytable==3.10.0
promise==2.3
prompt-toolkit==3.0.43
protobuf==3.20.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
pyzmq==25.1.2
readchar==4.0.5
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
runs==1.2.2
scann==1.2.10
scikit-surprise==1.1.3
scipy==1.12.0
six==1.16.0
stack-data==0.6.3
surprise==0.1
tdqm==0.0.1
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.1
tensorflow-datasets==4.9.4
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.36.0
tensorflow-metadata==1.14.0
tensorflow-recommenders==0.7.3
termcolor==2.4.0
toml==0.10.2
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
typing_extensions==4.5.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==3.0.1
wrapt==1.16.0
xmod==1.8.1
zipp==3.17.0
